LayerProvider.php as a new extension API point
